parse on paste and drop
